Implement gamma Test in the generator for Angel Benitez proposal
First implementation of the test. A single isotropic gamma of fixed energy is emitted
in the even event when the proper flag is set (/ActarSim/gun/gammaGenTest on ). As the
code is not ready for implantation (aborts if beam stops before vertex) the vertex point
is forced before the stopping of the beam (to be solved). No proton is still included for
the second branch of the 46Cr decay (46Cr->45V+p+gamma).