Add implantation mode and energy dispersion for the beam
Includes a implantation and decay mode which waits for the beam to stop before the reaction section.
Additionally a energy dispersion is included into the realisticBeam mode, with flat or gauss functions
to spread the energy around the incident energy value. Proper functions to call and control these
modes have been implemented. A technical problem in the call to GenerateBeam() and GeneratePrimaries()
has been solved modifiying the point where these functions are called, as well as removing unused
arguments in the function GenerateBeam().
The variable nextZVertex is modified (to 10^5) to ensure that never happens by default and remove
from the streamer to file (beamInfo).
A new macro implantAndDecay.mac has been introduced to exemplify the case.